requirements.txt
streamlit==1.27.2
opencv-python-headless
streamlit-js-eval
Pillow
numpy
torch
torchvision

